chore: decent loading API
