+++ crates/ratchet-core/src/tensor.rs
@@ -4,7 +4,7 @@ use crate::{
     RawStorage, Shape, Storage, Strides, TensorDType, TensorId,
 };
 use crate::{BinaryOp, LazyOp};
-use bytemuck::NoUninit;
+
 use derive_new::new;
 use parking_lot::RwLock;
 use std::sync::Arc;
