+++ crates/ratchet-models/Cargo.toml
@@ -14,3 +14,6 @@ thiserror.workspace = true
 derive-new.workspace = true
 log.workspace = true
 
+[dev-dependencies]
+hf-hub = { version = "0.3.0" }
+
