+++ crates/ratchet-models/src/lib.rs
@@ -1,1 +1,3 @@
 mod whisper;
+
+pub use whisper::*;
